Reworks tests for client
